Force overwrite executable when there is a dependency conflict (#1)
* Add task to remove rdkafka

At some point, the fluent-plugin-splunk-hec plugin failed to install due to a conflict of its json-jwt dependency with rdkafka. Here, we remove the latter dependency in order to allow the plugin installation to succeed.

* Set force boolean to true when installing Fluentd plugins

Set force flag to true on Fluentd plugin installation in order to overwrite conflicts with executable.